+++ src/main/java/live/smoothing/actuator/checker/ConditionChecker.java
@@ -1,0 +1,9 @@
+package live.smoothing.actuator.checker;
+
+import live.smoothing.actuator.config.ConditionSettings;
+import live.smoothing.actuator.dto.DataDTO;
+
+public interface ConditionChecker {
+
+    boolean checkCondition(DataDTO data, ConditionSettings.DeviceCondition settings);
+}
